Fix considering the comments
