--- src/utils/hooks.ts
@@ -1,12 +1,0 @@
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import { RootState, AppDispatch } from 'store';
-import { isEmptyObject } from './isEmptyObject';
-
-// Use throughout your app instead of plain `useDispatch` and `useSelector`
-export const useAppDispatch: () => AppDispatch = useDispatch;
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
-
-export const useAuth = () => {
-    const user = useSelector((state: RootState) => state.userReducer.user);
-    return user && !isEmptyObject(user);
-};
