+++ src/utils/hooks/index.ts
@@ -1,0 +1,11 @@
+import { useAppDispatch } from './useAppDispatch';
+import { useAppSelector } from './useAppSelector';
+import { useAuth } from './useAuth';
+import { useRedirectIfAuthenticated } from './useRedirectIfAuthenticated';
+
+export {
+    useAppDispatch,
+    useAppSelector,
+    useAuth,
+    useRedirectIfAuthenticated,
+};
