+++ src/utils/hooks/useRedirectIfAuthenticated.ts
@@ -1,0 +1,15 @@
+import { useEffect } from 'react';
+import { useNavigate } from 'react-router';
+import { routes } from 'Root';
+import { useAuth } from './useAuth';
+
+export const useRedirectIfAuthenticated = (redirectUrl = routes.main) => {
+    const isAuthenticated = useAuth();
+    const navigate = useNavigate();
+
+    useEffect(() => {
+        if (isAuthenticated) {
+            navigate(redirectUrl);
+        }
+    }, [isAuthenticated]);
+};
