--- src/utils/isEmptyObject.ts
@@ -1,1 +1,0 @@
-export const isEmptyObject = (obj: Record<string, unknown>) => Object.keys(obj).length === 0;
